day 3
code looks a bit too big, but it works. O(n) solution

a simple change that could improve a lot the memory usage (from O(n) to
O(1)) would be to, instead of storing the whole grid in memory, store
only a 3 line buffer. This would be enough to analyse the current line
for numbers and the two surrounding ones (above and below) for symbols.