day 14 part 2, phew!
realized that it is very likely that the board pattern cycles after some
executions.
So transformed the algorithm to execute for a while looking for
cycles (using Brent's tortoise and hare algorithm). Found the cycle, we
just extrapolate the result in the billion cycle.

Slowest solution so far, but couldn't think of a faster method.