day 20 solution and visualization!
got stuck for months in 20. Until I stopped thinking about the code and
started to look to the input. When plotting it as a graph, a clear
pattern became evident and I could finally see the beauty of the
circuit. No code was needed, apart from a simple lcm computation. Was
good to think out of (my) box at least once!