day05
tried to do a topological sort of the graph, but found cycles in the
official input. Bit weird, maybe not all nodes are used in the problems?

In the end, found a better solution that doesn't require inspecting all
n nodes, but is still using o(m^2) in part b (where `m` is the size of
the list to be checked).